# api/openapiv2/v1-tags/index-service/apis.swagger.yaml
swagger: "2.0"
info:
  title: chorus
  version: "1.0"
  contact:
    name: chorus index service
    url: https://github.com/CHORUS-TRE/chorus-backend
    email: dev@chorus-tre.ch
tags:
  - name: IndexService
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/v1/hello:
    get:
      summary: Get a hello
      description: This endpoint returns a hello
      operationId: IndexService_GetHello
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/chorusGetHelloReply'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      tags:
        - Index
      x-openapi-router-controller: index_controller
  /api/v1/hello/{identifier}:
    post:
      summary: Get a hello
      description: This endpoint returns a hello
      operationId: IndexService_CreateHello
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/chorusCreateHelloReply'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: identifier
          in: path
          required: true
          type: integer
          format: int32
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/IndexServiceCreateHelloBody'
      tags:
        - Index
      x-openapi-router-controller: index_controller
  /api/v1/helloo:
    get:
      summary: Get a hello
      description: This endpoint returns a hello
      operationId: IndexService_GetHelloo
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/chorusGetHelloReply'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      tags:
        - Index
      x-openapi-router-controller: index_controller
definitions:
  IndexServiceCreateHelloBody:
    type: object
    properties:
      title:
        type: string
      content:
        type: string
  chorusCreateHelloReply:
    type: object
    properties:
      identifier:
        type: integer
        format: int32
      title:
        type: string
      content:
        type: string
  chorusGetHelloReply:
    type: object
    properties:
      content:
        type: string
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
  - {}